+++ test-scripts-extra/read-asset-place.lua
@@ -1,0 +1,3 @@
+local game = remodel.readPlaceAsset("2983428072")
+
+assert(game.ClassName == "DataModel")