+++ src/remodel_context.rs
@@ -12,25 +12,35 @@ use rlua::{Context, UserData};
 #[derive(Clone)]
 pub struct RemodelContext {
     pub master_tree: Arc<Mutex<RbxTree>>,
+    auth_cookie: Option<String>,
 }
 
 impl RemodelContext {
-    pub fn get<'lua>(context: Context<'lua>) -> rlua::Result<Self> {
-        context.named_registry_value("remodel_context")
-    }
-
-    pub fn inject<'lua>(context: Context<'lua>) -> rlua::Result<()> {
+    pub fn new(auth_cookie: Option<String>) -> Self {
         let master_tree = Arc::new(Mutex::new(RbxTree::new(RbxInstanceProperties {
             name: "REMODEL ROOT".to_owned(),
             class_name: "REMODEL ROOT".to_owned(),
             properties: HashMap::new(),
         })));
 
-        let remodel_context = RemodelContext { master_tree };
+        Self {
+            master_tree,
+            auth_cookie,
+        }
+    }
 
-        context.set_named_registry_value("remodel_context", remodel_context)?;
+    pub fn get<'lua>(context: Context<'lua>) -> rlua::Result<Self> {
+        context.named_registry_value("remodel_context")
+    }
+
+    pub fn inject<'lua>(self, context: Context<'lua>) -> rlua::Result<()> {
+        context.set_named_registry_value("remodel_context", self)?;
 
         Ok(())
+    }
+
+    pub fn auth_cookie(&self) -> Option<&str> {
+        self.auth_cookie.as_ref().map(|v| v.as_str())
     }
 }